api/serializers.py
from rest_framework import serializers

from blog.models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'name', 'quantity', 'amount', 'description', 'created']