requirements.txt
django
psycopg[binary]
djangorestframework
environs